+++ torch/nn/_functions/activation.py
@@ -7,11 +7,10 @@ class Softsign(Function):
     def forward(ctx, input):
         ctx.save_for_backward(input)
         buffer = input.clone().abs_().add_(1)
-        return input.clone().div_(buffer)
+        return input.div(buffer)
 
     @staticmethod
     def backward(ctx, grad_output):
         input, = ctx.saved_variables
         buffer = input.abs().add_(1)
-        grad_input = grad_output.div(buffer.mul(buffer))
-        return grad_input
+        return grad_output.div(buffer.mul(buffer))
